refactor : List<Section> sections를 일급 컬렉션 Sections로 변경 (피드백 반영)
