+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,21 +1,10 @@
 package nextstep.subway.line.domain;
 
 import nextstep.subway.common.BaseEntity;
-import nextstep.subway.line.exception.CannotRemoveSectionException;
 import nextstep.subway.station.domain.Station;
-import nextstep.subway.station.exception.NotMatchingStationException;
-import nextstep.subway.station.exception.NotRemoveStationException;
-import nextstep.subway.station.exception.StationAlreadyExistException;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import static nextstep.subway.line.exception.LineExceptionMessage.*;
-import static nextstep.subway.station.exception.StationExceptionMessage.*;
 
 @Entity
 public class Line extends BaseEntity {
@@ -29,8 +18,8 @@ public class Line extends BaseEntity {
 
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections = new Sections();
 
     public Line() {
     }
@@ -46,48 +35,19 @@ public class Line extends BaseEntity {
     }
 
     public void addSection(Station upStation, Station downStation, int distance) {
-        if (sections.isEmpty()) {
-            sections.add(createSection(upStation, downStation, distance));
-            return;
-        }
-
-        validateUpStationMatching(upStation);
-        validateExistDownStation(downStation);
-
-        sections.add(createSection(upStation, downStation, distance));
-    }
-
-    private Section createSection(Station upStation, Station downStation, int distance) {
-        return new Section(this, upStation, downStation, distance);
-    }
-
-    private void validateUpStationMatching(Station upStation) {
-        if (!getLastDownStation().equals(upStation)) {
-            throw new NotMatchingStationException(EXCEPTION_MESSAGE_NOT_MATCHING_EXISTING_AND_NEW_STATION);
-        }
-    }
-
-    private void validateExistDownStation(Station downStation) {
-        if (sections.contains(downStation)) {
-            throw new StationAlreadyExistException(EXCEPTION_MESSAGE_EXIST_DOWN_STATION);
-        }
+        sections.addSection(this, upStation, downStation, distance);
     }
 
     public void deleteLastDownStation(Long downStation) {
-        Station lastDownStation = getLastDownStation();
-        validateDeletableStation(downStation, lastDownStation.getId());
-        sections.remove(lastDownStation);
+        sections.deleteLastDownStation(downStation);
     }
 
-    private void validateDeletableStation(Long downStation, Long lastDownStationId) {
-        boolean isEqualStation = lastDownStationId.equals(downStation);
+    public List<Section> getSections() {
+        return sections.getSections();
+    }
 
-        if (!isEqualStation) {
-            throw new NotRemoveStationException(EXCEPTION_MESSAGE_NOT_DELETABLE_STATION);
-        }
-        if (sections.size() == 1) {
-            throw new CannotRemoveSectionException(EXCEPTION_MESSAGE_NOT_DELETABLE_SECTION);
-        }
+    public List<Station> getStations() {
+        return sections.getStations();
     }
 
     public Long getId() {
@@ -102,20 +62,4 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public Station getLastDownStation() {
-        Section section = sections.get(sections.size() - 1);
-        return section.getDownStation();
-    }
-
-    public List<Section> getSections() {
-        return Collections.unmodifiableList(sections);
-    }
-
-    public List<Station> getStations() {
-        return sections.stream()
-                .sorted()
-                .flatMap(section -> Stream.of(section.getUpStation(), section.getDownStation()))
-                .distinct()
-                .collect(Collectors.toList());
-    }
 }
